Move the SpanMap definition into the span crate
