+++ crates/base-db/Cargo.toml
@@ -23,6 +23,5 @@ profile.workspace = true
 stdx.workspace = true
 syntax.workspace = true
 test-utils.workspace = true
-tt.workspace = true
 vfs.workspace = true
 span.workspace = true
